@@ -8,61 +8,72 @@
   "repository": "https://github.com/ruilisi/fortune-sheet",
   "license": "MIT",
   "scripts": {
-    "dev": "yarn storybook",
+    "dev": "npx yarn storybook",
     "build": "node ./scripts/build",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
     "tsc": "tsc --noEmit",
-    "lint": "eslint --cache --ext .js,.jsx,.ts,.tsx ./packages",
+    "lint": "eslint --cache --ext .js,.jsx,.ts,.tsx,.json ./packages",
     "postinstall": "lerna run build --scope=@fortune-sheet/formula-parser",
     "publish": "npm publish packages/core --registry 'https://registry.npmjs.org' && npm publish packages/react --registry 'https://registry.npmjs.org'",
     "test": "jest",
     "coverage": "jest --coverage"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.5",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
-    "@babel/preset-typescript": "^7.16.7",
-    "@storybook/addon-actions": "^6.4.19",
-    "@storybook/addon-essentials": "^6.4.19",
-    "@storybook/addon-interactions": "^6.4.19",
-    "@storybook/addon-links": "^6.4.19",
-    "@storybook/react": "^6.4.19",
-    "@storybook/testing-library": "^0.0.9",
-    "@testing-library/react": "^12.1.5",
-    "@types/jest": "^27.4.1",
-    "@typescript-eslint/eslint-plugin": "^5.13.0",
-    "@typescript-eslint/parser": "^5.13.0",
-    "babel-jest": "^27.5.1",
-    "babel-loader": "^8.2.3",
-    "eslint": "^8.10.0",
+    "@babel/core": "^7.20.7",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.18.6",
+    "@storybook/addon-actions": "^6.5.15",
+    "@storybook/addon-essentials": "^6.5.15",
+    "@storybook/addon-interactions": "^6.5.15",
+    "@storybook/addon-links": "^6.5.15",
+    "@storybook/addon-postcss": "^2.0.0",
+    "@storybook/react": "^6.5.15",
+    "@storybook/testing-library": "^0.0.13",
+    "@testing-library/react": "^13.4.0",
+    "@types/jest": "^29.2.5",
+    "@typescript-eslint/eslint-plugin": "^5.48.0",
+    "@typescript-eslint/parser": "^5.48.0",
+    "babel-jest": "^29.3.1",
+    "babel-loader": "^9.1.0",
+    "eslint": "^8.31.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.25.3",
-    "eslint-plugin-jest": "^26.1.3",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^27.1.7",
+    "eslint-plugin-jsonc": "^2.6.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "express": "^4.17.3",
     "father-build": "^1.21.2",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^27.5.1",
-    "jest-canvas-mock": "^2.3.1",
+    "jest": "^29.3.1",
+    "jest-canvas-mock": "^2.4.0",
+    "jsonc-eslint-parser": "^2.1.0",
     "lerna": "^4.0.0",
-    "lint-staged": "^12.3.4",
+    "lint-staged": "^13.1.0",
     "mongodb": "^4.5.0",
     "prettier": "^2.5.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "tsconfig-paths-webpack-plugin": "^3.5.2",
-    "typescript": "^4.6.2",
-    "ws": "^8.5.0",
+    "tsconfig-paths-webpack-plugin": "^4.0.0",
+    "ws": "^8.11.0",
     "yorkie": "^2.0.0"
   },
-  "dependencies": {},
+  "dependencies": {
+    "react": ">= 18.2.0",
+    "react-dom": "^18.2.0",
+    "typescript": "^4.9.4"
+  },
   "gitHooks": {
-    "pre-commit": "yarn lint"
+    "pre-commit": "npx yarn lint"
+  },
+  "resolutions": {
+    "react": "^18.2.0",
+    "eslint": "^8.31.0"
+  },
+  "peerDependencies": {
+    "react": ">= 18.2.0"
   }
 }
