@@ -18,8 +18,8 @@
     "lodash": "^4.17.21"
   },
   "peerDependencies": {
-    "react": ">= 16.8",
-    "react-dom": ">= 16.8"
+    "react": ">= 18.2",
+    "react-dom": ">= 18.2"
   },
   "keywords": [
     "spreadsheet",
