@@ -1322,7 +1322,7 @@ export function execFunctionGroup(
     updateValueOjects[key] = 1;
   } else {
     for (let x = 0; x < ctx.formulaCache.execFunctionExist.length; x += 1) {
-      const cell = ctx.formulaCache.execFunctionExist[x];
+      const cell = ctx.formulaCache.execFunctionExist[x] as any;
       const key = `r${cell.r}c${cell.c}i${cell.i}`;
       updateValueOjects[key] = 1;
     }
