@@ -1,10 +1,10 @@
 import React, { useRef } from "react";
 import { useOutsideClick } from "../../hooks/useOutsideClick";
 
-const MoreItemsContaier: React.FC<{ onClose?: () => void }> = ({
-  onClose,
-  children,
-}) => {
+const MoreItemsContaier: React.FC<{
+  onClose?: () => void;
+  children?: React.ReactNode;
+}> = ({ onClose, children }) => {
   const containerRef = useRef<HTMLDivElement>(null);
   useOutsideClick(
     containerRef,
